Mt 32 auth

+++ src/main/java/org/example/microTech/controllers/ProductController.java
@@ -1,9 +1,14 @@
 package org.example.microTech.controllers;
 
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.example.microTech.dto.*;
+import org.example.microTech.entities.User;
+import org.example.microTech.enums.UserRole;
+import org.example.microTech.exceptions.ForbiddenException;
+import org.example.microTech.exceptions.UnauthorizedException;
 import org.example.microTech.repositories.ProductRepository;
 import org.example.microTech.services.ProductService;
 import org.springframework.http.HttpStatus;
@@ -22,7 +27,10 @@ public class ProductController {
     private final ProductService productService;
 
     @PostMapping
-    public ResponseEntity<ApiResponse> createProduct(@Valid @RequestBody ProductRequestDTO dto) {
+    public ResponseEntity<ApiResponse> createProduct(@Valid @RequestBody ProductRequestDTO dto, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ProductResponseDTO product = productService.createProduct(dto);
 
         ApiResponse response =  ApiResponse.builder().message("Product created successfully!")
@@ -33,7 +41,10 @@ public class ProductController {
     }
 
     @GetMapping
-    public ResponseEntity<ApiResponse> getAllProducts() {
+    public ResponseEntity<ApiResponse> getAllProducts(HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         List<ProductResponseDTO> products = productService.getAllProducts();
         ApiResponse response =  ApiResponse.builder().message("All products")
                 .data(products)
@@ -46,8 +57,6 @@ public class ProductController {
     public ResponseEntity<ApiResponse> getProductById(@PathVariable Long id) {
         ProductResponseDTO product = productService.getProductById(id);
 
-
-
         ApiResponse response = ApiResponse.builder()
                 .message("Product found")
                 .data(product)
@@ -58,7 +67,10 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse> updateProduct(@PathVariable Long id, @Valid @RequestBody ProductRequestDTO dto) {
+    public ResponseEntity<ApiResponse> updateProduct(@PathVariable Long id, @Valid @RequestBody ProductRequestDTO dto, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ProductResponseDTO product = productService.updateProduct(id, dto);
 
         ApiResponse response = ApiResponse.builder()
@@ -71,7 +83,10 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-        public ResponseEntity<ApiResponse> deleteProduct(@PathVariable Long id) {
+        public ResponseEntity<ApiResponse> deleteProduct(@PathVariable Long id,HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ProductDeleteResponseDTO client = productService.deleteProduct(id);
         ApiResponse response = ApiResponse.builder()
                 .message("Supplier deleted successfully!")
